clean up the test case
